+++ CairoMakie/src/overrides.jl
@@ -320,6 +320,11 @@ function draw_plot(scene::Scene, screen::Screen,
         if basecolor isa Cairo.CairoPattern
             pattern_set_matrix(basecolor, Cairo.CairoMatrix(1, 0, 0, 1, 0, 0))
         end
+
+        for p in band.plots
+            p isa Mesh && continue
+            draw_plot(scene, screen, p)
+        end
     else
         for p in band.plots
             draw_plot(scene, screen, p)